+++ Api/TooltipDataApi.lua
@@ -23,6 +23,9 @@ function KrowiAF.AddTooltipDatum(achievementId, criteriaIndex, objectType, objec
 end
 
 local function AddTooltipData(achievementId, properties, criteria)
+    -- if not addon.Util.IsTable(criteria) then
+    --     criteria = {{0, criteria}};
+    -- end
     for _, criterium in next, criteria do
         KrowiAF.AddTooltipDatum(achievementId, criterium[1], criterium[3] or properties.ObjectType, criterium[2], criterium[4] or properties.Faction);
     end