+++ DataAddons/Retail/08_BattleForAzeroth/TooltipData.lua
@@ -30,6 +30,41 @@ local battleForAzerothMenuCriteria =  {
 };
 
 KrowiAF.TooltipData.BattleForAzeroth = {
+    { -- Raiding with Leashes V: Cuteaclysm
+        12079,
+        {
+            {1, 42178, type.Unit}, -- Magmatron
+            {1, 127850, type.Item}, -- Tinytron
+            {2, 41378, type.Unit}, -- Maloriak
+            {2, 127852, type.Item}, -- Discarded Experiment
+            {3, 41376, type.Unit}, -- Nefarian
+            {3, 127853, type.Item}, -- Rattlejaw
+            {4, 45992, type.Unit}, -- Valiona
+            {4, 127857, type.Item}, -- Twilight Clutch Sister
+            {5, 43735, type.Unit}, -- Elementium Monstrosity
+            {5, 127858, type.Item}, -- Bound Stream
+            {6, 43324, type.Unit}, -- Chogall
+            {6, 127859, type.Item}, -- Faceless Minion
+            {7, 45871, type.Unit}, -- Nezir
+            {7, 127863, type.Item}, -- Drafty
+            {8, 46753, type.Unit}, -- Alakir
+            {8, 127862, type.Item}, -- Zephyrian Prince
+            {9, 53691, type.Unit}, -- Shannox
+            {9, 127947, type.Item}, -- Blazehound
+            {10, 52498, type.Unit}, -- Bethtilac
+            {10, 127948, type.Item}, -- Cinderweb Recluse
+            {11, 53494, type.Unit}, -- Baleroc
+            {11, 127950, type.Item}, -- Surger
+            {12, 52571, type.Unit}, -- Majordomo Staghelm
+            {12, 127951, type.Item}, -- Infernal Pyreclaw
+            {13, 55312, type.Unit}, -- Yorsahj The Unsleeping
+            {13, 127952, type.Item}, -- Faceless Mindlasher
+            {14, 53879, type.Unit}, -- Deathwing
+            {14, 127953, type.Item}, -- Corrupted Blood
+            {15, 53879, type.Unit}, -- Deathwing
+            {15, 127954, type.Item}, -- Unstable Tendril
+        }
+    },
     { -- The Zandalari Menu
         12744,
         {
@@ -43,5 +78,112 @@ KrowiAF.TooltipData.BattleForAzeroth = {
             Faction = PLAYER_FACTION_GROUP.Horde,
         },
         battleForAzerothMenuCriteria
+    },
+    { -- Battle Safari
+        12930, type.Unit,
+        {
+            {1, 143031}, -- Shadowback Crawler
+            {2, 143032}, -- River Frog
+            {3, 143033}, -- Freshwater Crawler
+            {4, 143034}, -- Vale Marmot
+            {5, 143035}, -- Valley Chicken
+            {6, 143036}, -- Sandyback Crawler
+            {7, 143037}, -- River Otter
+            {8, 143039}, -- Parasitic Boarfly
+            {9, 143040}, -- Shack Crab
+            {10, 143041}, -- Inland Croaker
+            {11, 143042}, -- Giant Woodworm
+            {12, 143043}, -- Shore Butterfly
+            {13, 143044}, -- Barrier Hermit
+            {14, 143045}, -- Coastal Scuttler
+            {15, 143046}, -- Golden Beetle
+            {16, 143047}, -- Bloodfever Tarantula
+            {17, 143048}, -- Elusive Skimmer
+            {18, 143049}, -- Leafy Flutterwing
+            {19, 143050}, -- Young Sand Sifter
+            {20, 143051}, -- Sticky Oozeling
+            {21, 143052}, -- Returned Hatchling
+            {22, 143053}, -- Glutted Bleeder
+            {23, 143054}, -- Spectral Raven
+            {24, 143055}, -- Boghopper
+            {25, 143056}, -- Hermit Crab
+            {26, 143057}, -- Coastal Bounder
+            {27, 143038}, -- Honey Bee
+        }
+    },
+    { -- Raiding with Leashes VI: Pets of Pandaria
+        13469,
+        {
+            {1, 59915, type.Unit}, -- Jasper Guardian
+            {1, 60043, type.Unit}, -- Jade Guardian
+            {1, 60047, type.Unit}, -- Amethyst Guardian
+            {1, 60051, type.Unit}, -- Cobalt Guardian
+            {1, 150354, type.Item}, -- Stoneclaw
+            {2, 60143, type.Unit}, -- Garajal The Spiritbinder
+            {2, 150356, type.Item}, -- Wayward Spirit
+            {3, 60410, type.Unit}, -- Elegon
+            {3, 150357, type.Item}, -- Comet
+            {4, 60399, type.Unit}, -- Qin Xi
+            {4, 150360, type.Item}, -- Baoh Xi
+            {5, 62442, type.Unit}, -- Tsulong
+            {5, 150365, type.Item}, -- Azure Windseeker
+            {6, 62983, type.Unit}, -- Lei Shi
+            {6, 150372, type.Item}, -- Spirit Of The Spring
+            {7, 62980, type.Unit}, -- Imperial Vizier Zorlok
+            {7, 150374, type.Item}, -- Korthik Swarmling
+            {8, 62543, type.Unit}, -- Blade Lord Tayak
+            {8, 150375, type.Item}, -- Amberglow Stinger
+            {9, 62164, type.Unit}, -- Garalon
+            {9, 150377, type.Item}, -- Spawn Of Garalon
+            {10, 62511, type.Unit}, -- Amber Shaper Unsok
+            {10, 150380, type.Item}, -- Living Amber
+            {11, 62837, type.Unit}, -- Grand Empress Shekzeer
+            {11, 150381, type.Item}, -- Ravenous Prideling
+        }
+    },
+    { -- Mecha-Safari
+        13693, type.Unit,
+        {
+            {1, 154771}, -- Junkheap Roach
+            {2, 154775}, -- Fleeting Frog
+            {3, 154785}, -- Mechagon Marmot
+            {4, 154779}, -- Motorized Croaker
+            {5, 154791}, -- Scrapyard Tunneler
+            {6, 154769}, -- Duskytooth Snooter
+            {7, 154783}, -- Rustbolt Clucker
+            {8, 154767}, -- Rustyroot Snooter
+            {9, 154777}, -- Yellow Junkhopper
+            {10, 154773}, -- Experimental Roach
+            {11, 154787}, -- Specimen 97
+            {12, 154798}, -- Malfunctioning Microbot
+        }
+    },
+    { -- Nazjatari Safari
+        13694, type.Unit,
+        {
+            {1, 154716}, -- Spireshell Snail
+            {2, 154724}, -- Muck Slug
+            {3, 154702}, -- Sandclaw Sunshell
+            {4, 154814}, -- Abyssal Slitherling
+            {5, 154706}, -- Chitterspine Skitterling
+            {6, 154712}, -- Deeptide Fingerling
+            {7, 154697}, -- Sandclaw Pincher
+            {8, 154704}, -- Glimmershell Scuttler
+            {9, 154708}, -- Hissing Chitterspine
+            {10, 154710}, -- Great Sea Albatross
+            {11, 154714}, -- Bloodseeker
+        }
+    },
+    { -- From The Belly Of The Jelly
+        13715, type.Unit,
+        {
+            {1, 151651}, -- Slimy Darkhunter
+            {2, 151697}, -- Slimy Eel
+            {3, 151696}, -- Slimy Fangtooth
+            {4, 151632}, -- Slimy Hermit Crab
+            {5, 151673}, -- Slimy Octopode
+            {6, 151700}, -- Slimy Otter
+            {7, 151631}, -- Slimy Sea Slug
+        }
     },
 };