+++ Localization/ptBR.Plugins.lua
@@ -1,0 +1,9 @@
+local addonName, addon = ...;
+local L = LibStub(addon.Libs.AceLocale):NewLocale(addonName, "ptBR");
+if not L then return end
+
+-- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
+-- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
+
+-- [[ Exported at 2025-05-17 22-07-52 ]] --
+L["Add To Right Click Menu"] = "Adicionar ao Menu do Botão Direito"